readlong

+++ parcs/nnparcs/src/MAinMod.java
@@ -24,7 +24,7 @@ public class MAinMod {
         c.write(d);
 
         System.out.println("Waiting for result...");
-        System.out.println("Result: " + c.readLong());
+        System.out.println("Result: " + c.readInt());
         curtask.end();
 
 //        task curtask = new task();
